fix: allow feedback state push
