--- packages/dev-workflow-v2/domain-model/src/domain/ports/push-workflow-feature-branch.ts
@@ -1,5 +1,0 @@
-/**
- * @riviere-role domain-port
- * @riviere-role-justification MaintainerWorkflow pushes the feature branch recorded in workflow state through this capability.
- */
-export type PushWorkflowFeatureBranch = (branch: string) => void
