+++ packages/dev-workflow-v2/domain-model/src/domain/registry.ts
@@ -37,7 +37,6 @@ type WorkflowOperation =
   | 'record-ci-passed'
   | 'record-ci-failed'
   | 'create-pr'
-  | 'push-feedback-fixes'
   | 'verify-feedback-addressed'
 
 /** @riviere-role value-object */
