--- packages/dev-workflow-v2/use-cases/src/features/workflow/commands/push-feedback-fixes.ts
@@ -1,15 +1,0 @@
-import type { MaintainerWorkflow as Workflow } from '@living-architecture/dev-workflow-v2-domain-model/domain/workflow'
-import type { WorkflowCommandResult } from './workflow-command-result'
-
-/** @riviere-role command-use-case-input */
-export type PushFeedbackFixesInput = Record<never, never>
-
-/** @riviere-role command-use-case */
-export class PushFeedbackFixes {
-  constructor(private readonly workflow: Workflow) {}
-
-  execute(input: PushFeedbackFixesInput): WorkflowCommandResult {
-    void input
-    return { result: this.workflow.pushFeedbackFixes() }
-  }
-}
