+++ packages/dev-workflow-v2/use-cases/src/infra/external-clients/git/git-client.ts
@@ -40,15 +40,6 @@ export function readGitRepositoryStatus(
   }
 }
 
-/** @riviere-role external-client-service */
-export function pushGitBranch(
-  branch: string,
-  gitBinary = 'git',
-  executeGit: GitExecutor = defaultGitExecutor,
-): void {
-  runGit(executeGit, gitBinary, ['push', '-u', 'origin', branch])
-}
-
 function detectDefaultBranch(executeGit: GitExecutor, gitBinary: string): string {
   try {
     return runGit(executeGit, gitBinary, [
