+++ tools/dev-workflow-v2/.claude-plugin/plugin.json
@@ -1,5 +1,5 @@
 {
   "name": "dev-workflow-v2",
-  "version": "0.0.156",
+  "version": "0.0.157",
   "description": "Event-sourced workflow state machine for structured task lifecycle"
 }
