+++ tools/dev-workflow-v2/src/features/workflow/entrypoint/workflow/entrypoint.ts
@@ -5,7 +5,6 @@ import { RecordCiFailed } from '@living-architecture/dev-workflow-v2-use-cases/c
 import { RecordCiPassed } from '@living-architecture/dev-workflow-v2-use-cases/commands/record-ci-passed'
 import { RecordIssue } from '@living-architecture/dev-workflow-v2-use-cases/commands/record-issue'
 import { RecordPullRequest } from '@living-architecture/dev-workflow-v2-use-cases/commands/record-pull-request'
-import { PushFeedbackFixes } from '@living-architecture/dev-workflow-v2-use-cases/commands/push-feedback-fixes'
 import { VerifyFeedbackAddressed } from '@living-architecture/dev-workflow-v2-use-cases/commands/verify-feedback-addressed'
 import {
   parseNumberArgument,
@@ -40,7 +39,6 @@ export function createWorkflowRoutes(dependencies: CreateWorkflowRoutesEntrypoin
       }).result,
     createPullRequest: (workflow, args) =>
       new CreatePullRequest(workflow).execute({ arguments: args }).result,
-    pushFeedbackFixes: (workflow) => new PushFeedbackFixes(workflow).execute({}).result,
     recordCiPassed: (workflow) => new RecordCiPassed(workflow).execute({}).result,
     recordCiFailed: (workflow, output) => new RecordCiFailed(workflow).execute({ output }).result,
     verifyFeedbackAddressed: (workflow) => new VerifyFeedbackAddressed(workflow).execute({}).result,
