+++ tools/dev-workflow-v2/src/features/workflow/entrypoint/workflow/workflow-route-inputs.spec.ts
@@ -34,7 +34,6 @@ function buildWorkflow(
       prUrl: 'https://github.com/example/repo/pull/1',
       isDraft: false,
     }),
-    pushFeatureBranch: () => undefined,
     listSessionReviews: () => [],
     sleepMs: () => undefined,
     now: () => '2026-01-01T00:00:00Z',
