+++ tools/dev-workflow-v2/states/submitting_pr.md
@@ -10,7 +10,7 @@ You are creating a pull request.
 ## Constraints
 
 - `/dev-workflow-v2:create-pr` reads the recorded branch and issue, then delegates creation and recording to `workflow create-pr`.
-- Do not call `git push`; only `push-feedback-fixes` can push, and only while addressing feedback.
+- Do not call `git push`.
 - Do not call `gh pr create`, `gh pr edit`, `gh pr ready`, or `workflow record-pr` directly.
 - Cannot transition to AWAITING_CI unless prNumber is recorded
 - If blocked, transition to BLOCKED: `/dev-workflow-v2:workflow transition BLOCKED`
